고객정보 리스트

+++ src/main/java/com/datamon/datamon2/dto/repository/CustomerBasicConsultationDto.java
@@ -11,7 +11,7 @@ import java.time.Instant;
  */
 @Data
 public class CustomerBasicConsultationDto implements Serializable {
-    Integer idx;
+    Long idx;
     Long custId;
     String key;
     String value;
